Update README.md
This study uses Neural Networks as a tool to predict cancer development and/or possibility of recovery.  This will be investigated using the "UAE Cancer Patient Dataset." This dataset is designed for research, analysis, and machine learning applications in healthcare. It includes 10,000+ records of synthetic cancer patient data from the United Arab Emirates (UAE) with 20 features, such as:
✔ Patient demographics (Age, Gender, Nationality, Ethnicity)
✔ Diagnosis details (Cancer Type, Stage, Diagnosis Date)
✔ Treatment information (Treatment Type, Hospital, Physician)
✔ Health-related factors (Smoking Status, Comorbidities, Weight, Height)
✔ Outcomes (Recovered, Under Treatment, Deceased)1.2 Justification of the Topic